requirements.txt
ffmpeg
tk
pillow
wave
pydub
speechrecognition
fuzzywuzzy
python-Levenshtein